Take the output of DSpace MetadataExporter (a csv) and ZF Metadata (xml) from NFA and update the csv.
This is java. jbang is needed
It calls the xslt as `run.sh` would, but instead of creating a simple
archive format, it takes output of MetadataExporter and modifies that.
If we didn't care for mapped items (and their subjects) and if we would
not want a confirmation/summary of the changes done we could use the
traditional `bin/dspace import` with `--replace --extending-class
cz.cuni.mff.ufal.dspace.app.itemimport.ItemImportReplacingMetadata`